+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+### 05 Sept, 2019
+fix: show coins given correctly details.php  
+fix: undefined messages_class coins.php  
+revert: Adminer 4.7.2  
+fix: peers not adjusting correctly announce.php  
+
 ### 04 Sept, 2019
 update: reset auth remembered when confirming identity as staff  
 update: movies.php